Move some code from common to shared package
